Started diving into this project properly. Currently just finished step 4.
The next goal will be to scan the Pn biblio for files with Cr segs.Then for each file found, I will token the cr segs. Then I will parse those segs into a list: "the token before a comma should be the surname, the token after the comma+whitespace should be the journal abbreviation (most reviews are published in journals), the value of \d{4} (within parentheses) should be the date, and the page range will be hyphenated \d+-\d+ following p. Reviews in the online BMCRev (= 2293) will be an important class of edge case: these will have a hyperlink, not the other parseable strings."

This data will get put inot a class, including storing the given jounrals PN value, found by comparing against a csv file.

Then we're going to compare our list of Review entries that we gathered from the Pn files, to the list of reviews which should exist according to the CR element of each file, comparing on as many categories as possible. Compare the count and see how similar they are.

If they are complete matches, throw them away. Any incomplete matches need to be delt with.

if the only difference is a lakc of BP number in PN, update the file and move on. If PN has reviws with no BP token and BP doesn't have it create a list of those to send to BP.

if there are further differnces, print that output. and then gerate some xml from that

The file name and the value of idno[@type="pi"] should be the highest biblio number in the local directory +1