So I dropped in the new code files from the new version, tested that it runs in this folder.
A few notes:

First, some jouranl names were difficult to extract. If the program was stuggling too much with a name, it simply filled the jounral field as -1. It is thus necessary to check over the generated XML to make sure it has all the filled values. Most of the time it should be fine but good to double check.

Second, author names were a pain to deal with splitting and handling, so I don't. This means that running the name splitting tool on the generated content will be necessary.

Finally, I still need to go over the code and actually add in error handling.